chore: explicitly specify key in LogOverlay
React emits a warning if a key are missing in an array and falls back to
the index.

It is fine to use the index in this case because the index of a
particular element never changes, so explicitly state that by specifying
indices as keys and suppress the warnings.